feat(sales): provision + activate the sales-quote subsystem
The sales-quote tables (sales_quote / sales_reminder / sales_probability) were
retired from the legacy schema and exist in no database, so the Sales page could
only ever render empty. Per the product decision, re-provision and activate them.

- Self-provision the three tables at startup on Oracle (AbisSchema.OwnedDdl) with
  the authoritative legacy column names; ids are MAX+1 (added to Database:MaxIdTables
  — no legacy sequence exists). Idempotent (swallows ORA-00955). Canonical DDL in
  docs/data-model/migrations/004_sales_quote.sql. No-op on SQLite (fixture models them).
- Add the missing create-quote path: POST /api/sales/quotes (SalesQuoteWrite),
  customerId required + existence-checked (clean 400, not ORA-02291), a new quote is
  revision 1, created_date defaults to now. New "New quote" form on the Sales page
  (posts via authFetch, like the truck page — new endpoint, no client regen).
- Fix a latent bug the create-quote test surfaced: SalesQuoteListRow.LatestProbability
  was int? but is a MAX() aggregate (Int64 on SQLite, NUMBER on Oracle) — widened to
  long?, which was throwing an InvalidCastException whenever a quote had a probability.

OpenAPI snapshot regenerated. 302 tests pass, incl. a new create → round-trip test.

+++ api/src/ABIS.Api/Models/Entities.cs
@@ -904,7 +904,9 @@ public sealed class SalesQuoteListRow
     public decimal? TotalLbProcessed { get; set; }
     public DateTime? CreatedDate { get; set; }
     public DateTime? ValidDate { get; set; }
-    public int? LatestProbability { get; set; }
+    // long? (not int?): this is a MAX() aggregate, which SQLite returns as Int64 and Oracle as NUMBER —
+    // narrowing to int? throws an InvalidCastException as soon as a quote actually has a probability.
+    public long? LatestProbability { get; set; }
 }
 
 /// <summary>A sales quote header (table <c>sales_quote</c>, composite key
